Added Function to draw faces and an Engine Class
Face drawing function is temporary, ability to decipher which face is in front is not currently in. Engine class has objects such as EngineObject and EngineSettings. EngineObject currently only has a position and velocity component. EngineSettings currently is used to control refresh times on the three timers, update, renderUpdate, and physicsUpdate.